chore: code docs of server updated
