+++ server/helpers.go
@@ -5,8 +5,43 @@ import (
 	"fmt"
 	"log"
 	"strings"
+
+	"gorm.io/gorm"
 )
 
+// Migrate database table
+func (server *Server) MigrateDatabaseTables() {
+	err := server.DB_CLIENT.AutoMigrate(&Domain{})
+	if err != nil {
+		panic(err)
+	}
+	err = server.DB_CLIENT.AutoMigrate(&GitCredential{})
+	if err != nil {
+		panic(err)
+	}
+	err = server.DB_CLIENT.AutoMigrate(&ApplicationSource{})
+	if err != nil {
+		panic(err)
+	}
+	err = server.DB_CLIENT.AutoMigrate(&Application{})
+	if err != nil {
+		panic(err)
+	}
+	err = server.DB_CLIENT.AutoMigrate(&ApplicationBuildLog{})
+	if err != nil {
+		panic(err)
+	}
+	err = server.DB_CLIENT.AutoMigrate(&IngressRule{})
+	if err != nil {
+		panic(err)
+	}
+	err = server.DB_CLIENT.AutoMigrate(&RedirectRule{})
+	if err != nil {
+		panic(err)
+	}
+}
+
+// Generate Git Repository URL from ApplicationSource
 func (src ApplicationSource) RepositoryURL() string {
 	if src.GitProvider == GitProviderGithub {
 		return "https://github.com/" + src.RepositoryUsername + "/" + src.RepositoryName + ".git"
@@ -17,10 +52,8 @@ func (src ApplicationSource) RepositoryURL() string {
 	return ""
 }
 
+// Add a log to the application build log
 func (s *Server) AddLogToApplicationBuildLog(log_id string, message string, loglevel string, add_newline bool) {
-	// push all the logs to redis topic for realtime log streaming
-	// topic -> log_update/<log_id>
-
 	var logRecord ApplicationBuildLog
 	tx := s.DB_CLIENT.Where("id = ?", log_id).First(&logRecord)
 	if tx.Error != nil {
@@ -39,15 +72,19 @@ func (s *Server) AddLogToApplicationBuildLog(log_id string, message string, logl
 	} else if loglevel == "success" {
 		message = fmt.Sprintf("\x1b[1;32m\x1b[0m%s", message)
 	}
-
 	// don't need special modification for -> loglevel == "info" || loglevel == ""
+
+	// Appends the log in database
 	logRecord.Logs = logRecord.Logs + message
 	s.DB_CLIENT.Save(&logRecord)
 
-	// push to redis topic
+	// push the logs to redis topic for realtime log streaming
+	// topic -> log_update/<log_id>
 	s.REDIS_CLIENT.Publish(context.Background(), "log_update/"+log_id, message)
 }
 
+// Mark the application build log as completed
+// Marking the log as completed will stop the log streaming
 func (s *Server) MarkBuildLogAsCompleted(log_id string) {
 	var logRecord ApplicationBuildLog
 	tx := s.DB_CLIENT.Where("id = ?", log_id).First(&logRecord)
@@ -59,6 +96,8 @@ func (s *Server) MarkBuildLogAsCompleted(log_id string) {
 	s.DB_CLIENT.Save(&logRecord)
 }
 
+// Create a default git user if not exists
+// Username -> default & Password -> ""
 func (s *Server) CreateDefaultGitUser() {
 	var git_credential GitCredential
 	tx := s.DB_CLIENT.Where("name = ?", "default").First(&git_credential)
@@ -75,6 +114,12 @@ func (s *Server) CreateDefaultGitUser() {
 	}
 }
 
+/*
+Generate summary text of Application Code Source for displaying in the UI
+For Git -> Show Repository URL & Branch
+For Tarball -> Show `Source Code`
+For Image -> Show Docker Image Name
+*/
 func (s ApplicationSource) GetSourceSummary() string {
 	if s.Type == ApplicationSourceTypeGit {
 		return fmt.Sprintf("%s Branch: %s", s.RepositoryURL(), s.Branch)
@@ -88,6 +133,30 @@ func (s ApplicationSource) GetSourceSummary() string {
 	return "Unknown"
 }
 
+// Function to check if the server is running in production environment
 func (s *Server) isProductionEnvironment() bool {
 	return strings.Compare(s.ENVIRONMENT, "production") == 0
+}
+/*
+Function to set application status to `building_image_failed` and update in database
+Used when the image build fails
+*/
+func failImageBuildUpdateStatus(application *Application, db_client gorm.DB) {
+	application.Status = ApplicationStatusBuildingImageFailed
+	tx := db_client.Save(&application)
+	if tx.Error != nil {
+		log.Println("Failed to update application status in database")
+	}
+}
+
+/*
+Function to set application status to `deploying_failed` and update in database
+Used when the application deploy fails
+*/
+func failApplicationDeployUpdateStatus(application *Application, db_client gorm.DB) {
+	application.Status = ApplicationStatusDeployingFailed
+	tx := db_client.Save(&application)
+	if tx.Error != nil {
+		log.Println("Failed to update application status in database")
+	}
 }
