+++ server/utils.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 )
 
-// FetchRepositoryUsernameFromURL returns the username from a repository url.
+// Returns the owner username from a repository url.
 func FetchRepositoryUsernameFromURL(repo_url string) string {
 	// parse
 	url, err := url.Parse(repo_url)
@@ -21,7 +21,7 @@ func FetchRepositoryUsernameFromURL(repo_url string) string {
 	return ""
 }
 
-// FetchRepositoryNameFromURL returns the repository name from a repository url.
+// Returns the repository name from a repository url.
 func FetchRepositoryNameFromURL(repo_url string) string {
 	// parse
 	url, err := url.Parse(repo_url)
@@ -36,7 +36,7 @@ func FetchRepositoryNameFromURL(repo_url string) string {
 	return ""
 }
 
-// FetchGitProviderFromURL returns the git provider from a repository url.
+// Returns the git provider from a repository url.
 func FetchGitProviderFromURL(repo_url string) GitProvider {
 	if strings.Contains(repo_url, "github.com") {
 		return GitProviderGithub
@@ -48,7 +48,11 @@ func FetchGitProviderFromURL(repo_url string) GitProvider {
 	return ""
 }
 
-// Sanitize the fileName to remove potentially dangerous characters
+/*
+Sanitize the fileName to remove potentially dangerous characters
+It's meant to be used for filename
+Should not use this to sanitize file path
+*/
 func SanitizeFileName(fileName string) string {
 	// Remove any path components and keep only the file name
 	fileName = filepath.Base(fileName)
